added specs
