+++ app/controllers/concerns/event_tracking.rb
@@ -1,0 +1,20 @@
+module EventTracking
+  extend ActiveSupport::Concern
+
+  def index; end
+
+  def toggle
+    @analytics_config.value = params[:analytics]
+    if @analytics_config.save
+      redirect_to project_event_tracking_index_path, notice: notice_message
+    else
+      redirect_to project_event_tracking_index_path, alert: @analytics_config.errors.full_messages.join('; ')
+    end
+  end
+
+  protected
+
+  def notice_message
+    "Event tracking successfully #{params[:analytics] == 'true' ? 'enabled' : 'disabled'}!"
+  end
+end
